- Added directoryElement::removeElement - Bugfixes in unlinkElement (remove entry in parent directoryElement and check for tree root)

+++ trunk/mirror/rw_functions.cpp
@@ -10,16 +10,21 @@
 #include "rw_functions.h"
 
 bool unlinkElement(directoryElement* theElement, iNodeMap& nodeSet) {
+	if (theElement->get_parent() == NULL) return false;
+	
 	if (theElement->isDirectory()) {
 		for (list<directoryElement>::iterator it = theElement->get_contents()->begin();it!=theElement->get_contents()->end();it++) {
 			unlinkElement(&(*it),nodeSet);
 		}
 		theElement->get_contents()->clear();
-		rmdir((theElement->getPathToElement()).c_str());
-	} else unlink((theElement->getPathToElement()).c_str());
+	}
+	
+	remove((theElement->getPathToElement()).c_str());
 	
 	if (theElement->get_node()->remove_element(theElement) == 0)
 		nodeSet.deleteNode(theElement->get_node());
+	
+	theElement->get_parent()->remove_element(theElement);
 	
 	return true;
 }
